+++ previous-versions/sync-for-commerce-version-1/codatsynccommerce.go
@@ -197,9 +197,9 @@ func New(opts ...SDKOption) *CodatSyncCommerce {
 		sdkConfiguration: sdkConfiguration{
 			Language:          "go",
 			OpenAPIDocVersion: "1.1",
-			SDKVersion:        "0.23.0",
-			GenVersion:        "2.188.1",
-			UserAgent:         "speakeasy-sdk/go 0.23.0 2.188.1 1.1 github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1",
+			SDKVersion:        "0.23.1",
+			GenVersion:        "2.194.1",
+			UserAgent:         "speakeasy-sdk/go 0.23.1 2.194.1 1.1 github.com/codatio/client-sdk-go/previous-versions/sync-for-commerce-version-1",
 		},
 	}
 	for _, opt := range opts {